mv instrumentor to own file
