+++ src/main/python/alpacalert/models.py
@@ -3,11 +3,9 @@
 from __future__ import annotations
 
 import itertools
-import logging
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
 from enum import Enum
-from typing import Any, Iterable, TypeVar
+from typing import Iterable
 
 from pydantic import BaseModel
 
@@ -146,53 +144,3 @@ class Visualiser(ABC):
 
 def flatten(its: Iterable[Iterable]) -> list:
 	return list(itertools.chain.from_iterable(its))
-
-
-class Instrumentor:
-	@dataclass(frozen=True)
-	class Kind:
-		namespace: str
-		name: str
-
-	Registrations = Iterable[tuple[Kind, type["Instrumentor"]]]
-
-	@abstractmethod
-	def registrations(self) -> Instrumentor.Registrations: ...
-
-	@abstractmethod
-	def instrument(self, registry: InstrumentorRegistry, kind: Instrumentor.Kind, *arg, **kwargs) -> list[Scanner]: ...
-
-
-class InstrumentorRegistry:
-	"""
-	Instrument an external entity by generating Sensors, Systems, or Services.
-	"""
-
-	Registry = dict[Instrumentor.Kind, Instrumentor]
-
-	def __init__(self, instrumentors: Registry | None = None):
-		self.instrumentors: InstrumentorRegistry.Registry
-		if instrumentors:
-			self.instrumentors = instrumentors
-		else:
-			self.instrumentors = {}
-
-	def instrument(self, kind: Instrumentor.Kind, *args, **kwargs) -> list[Scanner]:
-		"""
-		Instrument an external entity by generating Sensors, Systems, or Services.
-		"""
-		instrumentor = self.instrumentors.get(kind)
-		if instrumentor:
-			try:
-				return instrumentor.instrument(self, kind, *args, **kwargs)
-			except Exception as e:
-				raise InstrumentorError(f"failed to instrument {kind=}") from e
-		else:
-			raise InstrumentorError(f"no provider {kind=}")
-
-	def register(self, kind: Instrumentor.Kind, instrumentor: Instrumentor):
-		self.instrumentors[kind] = instrumentor
-
-
-class InstrumentorError(Exception):
-	"""An error instrumenting an object"""
